+++ share/picongpu/examples/LaserWakefield/lib/python/picongpu/params.py
@@ -55,7 +55,7 @@ PARAMETERS = {
             default=1.0,
             range=(0.1, 10.0),
             pic_to_SI=lambda steps: steps * dt,
-            pic_from_SI=lambda time: int(round(time / dt)),
+            pic_from_SI=lambda time: round(time / dt),
             label="simulation time",
         )
     ],
